Struct spirit_hyper::BuildServer

pub struct BuildServer<BS>(pub BS);

A Transformation to turn a Builder into a Server.

The value wrapped in this shall be a closure taking:

The Builder

The configuration fragment (HyperServer)

A &str name

A Receiver<()>.

It shall produce a server wrapped with a graceful shutdown. Technically, it can produce whatever future that'll terminate once the Receiver contains something.

Constructing the server builder directly is a bit of a chore (bunch of cloning, lifetimes, and it's something like 3rd-order function ‒ function that returns a function that returns a function...). In many cases, the server_from_handler can help.

An instance of ServerBuilder goes inside.